fix: moving the code to a separate component for modal
